Share some helpers and fixtures
