+++ forklift/forklift/pipeline/entities/generic.py
@@ -1,0 +1,8 @@
+from dataclasses import dataclass
+from typing import Union
+
+
+@dataclass
+class IdRange:
+    id_min: Union[int, str]
+    id_max: Union[int, str]
